fixed more issues
